+++ README.md
@@ -4,18 +4,49 @@
      \__, |\___|_|  |_| |_| |_|___/
      |___/                               
 
-an opinionated hack-space for building things fast without doing a bunch of time setting things up
+an opinionated hack-space for building things fast without spending a bunch of time setting things up
 
-# Installation
+`germs` can be used in two different ways, either as a project template (there's some silly arithmetic and CLI stuff you can erase and replace with cool new stuff) or as a dependency to your module.
 
--   `git clone https://github.com/brekk/germs.git PROJECT_NAME` - pull in the project
--   `cd PROJECT_NAME` - go to the project
--   `npm i` or `yarn install` -- install dependencies for this project
-    -   you may need to `brew install yarn` first (on osx) -- otherwise you might want [this link](https://yarnpkg.com/en/docs/install) instead.
--   `npm i nps -g` or `yarn install nps global` -- add `nps` tool for better development QoL
--   `nps` - list the general commands of building
+## Installation
 
-# Commands
+Installing `germs`:
+
+`yarn install germs -D` or `npm i germs -D`
+
+Installing `nps` (a delightful build tool):
+
+`yarn install nps global` or `npm i nps -g`
+
+## Opinions
+
+`germs` relies on the following modules for a good build experience:
+
+### Build
+
+-   `nps` for build tasks
+
+### Compile
+
+-   `babel` for per-file conversion
+-   `rollup` for bundled conversion
+
+### Test
+
+-   `jest` for testing
+-   `execa` for testing CLI implementations
+-   `dont-break` for downstream testing
+
+### Quality of Life for Developers
+
+-   `eslint` for static analysis rules (static analysis rules!)
+-   `documentation` for automatic jsdoc > documentation generation
+-   `madge` for dependencies and graphs
+-   `depcheck` for (sometimes-unreliable) dependency checking
+-   `updtr` for automatic package updates
+-   `husky` for some commit conventions (use `--no-verify` to skip)
+
+## Commands
 
 -   `nps dependencies.check` - check dependencies
 -   `nps dependencies.graph` - generate a visual dependency graph
@@ -33,11 +64,83 @@ an opinionated hack-space for building things fast without doing a bunch of time
 -   `nps care` - run all the things
 -   `nps precommit` - nps care
 
-# Contributions welcome!
+#### Contributions welcome!
 
-# API
+#### Example API
 
 <!-- Generated by documentation.js. Update this documentation by updating the source code. -->
+
+##### add
+
+add things
+
+**Parameters**
+
+-   `a` **[number](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number)** a number
+-   `b` **[number](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number)** b number
+
+**Examples**
+
+```javascript
+import {add} from 'f-utility'
+add(4, 2) // 6
+```
+
+Returns **[number](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number)** sum
+
+##### divide
+
+divide things
+
+**Parameters**
+
+-   `a` **[number](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number)** a number
+-   `b` **[number](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number)** b number
+
+**Examples**
+
+```javascript
+import {divide} from 'f-utility'
+divide(4, 2) // 0.5
+```
+
+Returns **[number](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number)** divided
+
+##### multiply
+
+multiply things
+
+**Parameters**
+
+-   `a` **[number](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number)** a number
+-   `b` **[number](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number)** b number
+
+**Examples**
+
+```javascript
+import {multiply} from 'f-utility'
+multiply(4, 2) // 8
+```
+
+Returns **[number](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number)** multiplied
+
+##### subtract
+
+subtract things
+
+**Parameters**
+
+-   `a` **[number](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number)** a number
+-   `b` **[number](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number)** b number
+
+**Examples**
+
+```javascript
+import {subtract} from 'f-utility'
+subtract(4, 2) // -2
+```
+
+Returns **[number](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number)** subtracted
 
 ## add
