+++ package.json
@@ -1,10 +1,12 @@
 {
   "name": "germs",
-  "version": "0.0.1",
-  "description": "make some stuff",
-  "main": "germs.js",
+  "version": "0.0.0",
+  "description": "an opinionated build toolset that is dead-simple so you can focus on writing code",
+  "main": "lib/germs-builder.js",
   "scripts": {
     "dependencies": "nps dependencies",
+    "install": "npm run rebuild",
+    "rebuild": "babel src -d lib --ignore *.spec.js",
     "readme": "nps readme",
     "lint": "nps lint",
     "test": "nps test",
@@ -19,8 +21,8 @@
   },
   "keywords": [
     "build",
-    "root",
-    "base"
+    "code",
+    "simply"
   ],
   "author": "brekk",
   "license": "ISC",
@@ -54,6 +56,21 @@
   },
   "dependencies": {
     "katsu-curry": "^0.4.0"
+  },
+  "peerDependencies": {
+    "babel-cli": "^6.26.0",
+    "babel-core": "^6.26.0",
+    "babel-eslint": "^8.0.0",
+    "babel-jest": "^21.0.2",
+    "depcheck": "^0.6.7",
+    "documentation": "^5.3.1",
+    "eslint": "^4.6.1",
+    "husky": "^0.14.3",
+    "jayin": "^0.0.3",
+    "jest": "^21.0.2",
+    "nps-utils": "^1.3.0",
+    "nps": "^5.7.1",
+    "rollup": "^0.49.3"
   },
   "devDependencies": {
     "babel-cli": "^6.26.0",
